chore: migrate from Vite to Next.js app router
