+++ next.config.ts
@@ -1,0 +1,12 @@
+import type { NextConfig } from 'next'
+import path from 'path'
+
+const nextConfig: NextConfig = {
+  output: 'export',
+  distDir: 'build',
+  turbopack: {
+    root: path.resolve(__dirname),
+  },
+}
+
+export default nextConfig
