+++ package.json
@@ -4,8 +4,9 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "npx serve@latest ./build",
     "lint": "eslint .",
     "preview": "vite preview",
     "storybook": "storybook dev -p 5174",
@@ -29,6 +30,7 @@
     "classnames": "^2.5.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.525.0",
+    "next": "^16.2.5",
     "next-themes": "^0.4.6",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
@@ -47,6 +49,7 @@
     "@types/node": "^24.0.10",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
+    "@tailwindcss/postcss": "^4.1.11",
     "@vitejs/plugin-react-swc": "^3.10.2",
     "@vitest/browser": "^3.2.4",
     "@vitest/coverage-v8": "^3.2.4",
