+++ postcss.config.mjs
@@ -1,0 +1,1 @@
+export default { plugins: { '@tailwindcss/postcss': {} } }
