--- src/App.tsx
@@ -1,8 +1,0 @@
-import "./App.css";
-import { Home } from "./pages/home/home";
-
-function App() {
-  return <Home />;
-}
-
-export default App;
