+++ src/components/ui/avatar.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import * as React from "react";
 import * as AvatarPrimitive from "@radix-ui/react-avatar";
