+++ src/components/ui/checkbox.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import * as React from "react";
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox";
 import { CheckIcon } from "lucide-react";
