+++ src/components/ui/slider.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import * as React from "react";
 import * as SliderPrimitive from "@radix-ui/react-slider";
