+++ src/components/ui/tooltip.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import * as React from "react";
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
