--- src/pages/home/components.tsx
@@ -1,3 +1,0 @@
-export const Components = () => {
-  return <div></div>;
-};
